+++ lib/application/sign_up/sign_up_state.dart
@@ -5,22 +5,18 @@ class SignUpState with _$SignUpState {
   const factory SignUpState({
     required EmailAddress emailAddress,
     required Password password,
-    required PhoneNumber phone_number,
-    required bool showErrorMessages,
+    //required PhoneNumber phone_number,
     required bool isSubmitting,
     required bool isEmailVerified,
-    required bool isPhoneNumberVerified,
     required Option<Either<UserFailure, Unit>> userFailureOrUserSuccess,
   }) = _SignUpState;
 
   factory SignUpState.initial() => SignUpState(
         emailAddress: EmailAddress(''),
         password: Password(''),
-        phone_number: PhoneNumber(' '),
-        showErrorMessages: false,
+        //phone_number: PhoneNumber(' '),
         isSubmitting: false,
         isEmailVerified: false,
-        isPhoneNumberVerified: false,
         userFailureOrUserSuccess: none(),
       );
 }
