+++ lib/injection.config.dart
@@ -14,7 +14,7 @@ import 'package:get_it/get_it.dart' as _i174;
 import 'package:injectable/injectable.dart' as _i526;
 
 import 'application/sign_in/sign_in_bloc.dart' as _i939;
-import 'application/sing_up/sign_up_bloc.dart' as _i1063;
+import 'application/sign_up/sign_up_bloc.dart' as _i1011;
 import 'domain/user/interface_user_facade.dart' as _i746;
 import 'infrastructure/core/core_module.dart' as _i189;
 import 'infrastructure/firebase/user_firebase_repository.dart' as _i191;
@@ -39,10 +39,10 @@ extension GetItInjectableX on _i174.GetIt {
               gh<_i59.FirebaseAuth>(),
               gh<_i974.FirebaseFirestore>(),
             ));
+    gh.factory<_i1011.SignUpBloc>(
+        () => _i1011.SignUpBloc(gh<_i746.InterfaceUserFacade>()));
     gh.factory<_i939.SignInBloc>(
         () => _i939.SignInBloc(gh<_i746.InterfaceUserFacade>()));
-    gh.factory<_i1063.SignUpBloc>(
-        () => _i1063.SignUpBloc(gh<_i746.InterfaceUserFacade>()));
     return this;
   }
 }
